v8
Strong change to the structure of the Project. Move some functions to a
new file for optimization and management. Solved some big errors caused
by problem with variables
