+++ results.txt
@@ -9,3 +9,126 @@ Description: Music video by Nirvana performing Come As You Are. (C) 1992 Geffen 
 Title: Nirvana - Come As You Are
 date: Caricato il 16 giu 2009
 Views: 121.223.508
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.342.886
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.342.886
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.342.886
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.342.886
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Title: Nirvana - Come As You Are
+date: Caricato il 16 giu 2009
+Views: 121.343.489
+Average views per day: None
+
+ 
+Title: Nirvana - Lithium
+date: Caricato il 16 giu 2009
+Views: 64.698.890
+Average views per day: None
+
+ 
+Title: Nirvana - Smells Like Teen Spirit
+date: Caricato il 16 giu 2009
+Views: 332.998.869
+Average views per day: None
+
+ 
+Title: Nirvana - Heart-Shaped Box
+date: Caricato il 16 giu 2009
+Views: 74.818.363
+Average views per day: None
+
+ 
+Title: Nirvana - The Man Who Sold The World (MTV Unplugged)
+date: Caricato il 16 giu 2009
+Views: 80.265.358
+Average views per day: None
+
+ 
+Title: Red Hot Chili Peppers - Californication [Official Music Video]
+date: Caricato il 26 ott 2009
+Views: 205.538.256
+Average views per day: 90865.71883289125
+
+ 
+Title: Pink Floyd - Another Brick In The Wall (HQ)
+date: Caricato il 05 lug 2010
+Views: 111.023.371
+Average views per day: 55235.50796019901
+
+ 
+Title: The Cranberries - Zombie
+date: Caricato il 16 giu 2009
+Views: 313.157.527
+Average views per day: 130809.32623224729
+
+ 
+Title: Nirvana - In Bloom
+date: Caricato il 16 giu 2009
+Views: 49.023.087
+Average views per day: 20477.479949874687
+
+ 
+Title: Nirvana - Where Did You Sleep Last Night (Legendado)
+date: Pubblicato il 22 gen 2013
+Views: 2.096.948
+Average views per day: 1945.2207792207791
+
+ 
+Title: Guns N' Roses - November Rain
+date: Caricato il 24 dic 2009
+Views: 499.570.668
+Average views per day: 226768.34679981842
+
+ 
+Title: nirvana - rape me
+date: Caricato il 22 giu 2008
+Views: 4.632.765
+Average views per day: 1682.8060297856885
+
+ 
+Title: Guns N' Roses - Sweet Child O' Mine
+date: Caricato il 24 dic 2009
+Views: 255.998.400
+Average views per day: 1162.0444847934634
+
+ 
+Title: System Of A Down - Chop Suey!
+date: Caricato il 02 ott 2009
+Views: 358.904.105
+Average views per day: 157000.9208223972
+
+ 
+Title: Nickelback - How You Remind Me
+date: Caricato il 26 ott 2009
+Views: 154.369.422
+Average views per day: 68244.66047745358
+
+
